+++ crates/zng/src/lib.rs
@@ -447,11 +447,13 @@
 //! some catastrophic video driver errors, like a forced disconnect caused by a driver update. The [`task::spawn`] and related
 //! fire-and-forget task runners will also just log the panic as an error.
 //!
-//! You can also use the [`zng::app::crash_handler`] to collect panic backtraces, crash minidumps, show a crash dialog to the user
+//! The [`zng::app::crash_handler`] is enabled by default, it collect panic backtraces, crash minidumps, show a crash dialog to the user
 //! and restart the app. During development a debug crash dialog is provided, it shows the stdout/stderr, panics stacktrace and
-//! minidumps collected if any non-panic fatal error happens.
+//! minidumps collected if any non-panic fatal error happens. Note that the crash handler **stops debuggers from working**, see the
+//! [Debugger section] of the crash-handler docs on how to automatically disable the crash handler for debugger runs.
 //!
 //! [`tracing`]: https://docs.rs/tracing
+//! [Debugger section]: zng::app::crash_handler#debugger
 //!
 //! # In-Depth Documentation
 //!