+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Add `ZNG_NO_CRASH_HANDLER` env var to easily disable crash handler for special runs like for a debugger.
 * Add `CrashConfig::no_crash_handler` for custom crash handler disabling.
+* Add `zng::app::print_tracing_filter`.
 
 # 0.7.1